:art: Rétabli le bouton "déplier" dans les sommes...
Maintenant pour toutes les règles, y compris quand ce ne sont pas des sommes
